Fix compiler deprecation and exhaustivity warnings (#1244)
- Renderer.scala (x2), StartupRenderingSpec.scala: explicit .toFloat
  instead of the deprecated implicit Int-to-Float widening
- BenchmarkRunner.scala: Thread.threadId() instead of the deprecated
  Thread.getId()
- WorkspaceTreeLifecycleSpec.scala: collapse the non-Split match arms
  to a single wildcard case so the match is exhaustive against
  WorkspaceNode.DockedSurface

+++ src/test/scala/com/serenity/StartupRenderingSpec.scala
@@ -276,7 +276,7 @@ class StartupRenderingSpec extends AnyFlatSpec with Matchers:
       com.serenity.ui.layout.TextAreaPx(
         codeMetrics.toPixelX(contentRect.x).toFloat,
         codeMetrics.toPixelY(contentRect.y + contentRect.height / 2),
-        contentRect.width * codeMetrics.charWidth,
+        contentRect.width.toFloat * codeMetrics.charWidth,
         codeMetrics.lineHeight
       ),
       textFont,
